// src/simple/buttonApp/types.ts
export type AlignButtonType = 'center' | 'space-between'
export type VariantButtonType = 'contained' | 'outlined'
export type ColorButtonType = 'default' | 'success' | 'error'

export type KeyStyleColor =
  | 'default'
  | 'success'
  | 'error'
  | 'default_title'
  | 'success_title'
  | 'error_title'
  | 'default_subtitle'
  | 'success_subtitle'
  | 'error_subtitle'